rocket_sim/requirements.txt
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
numpy-stl>=3.0.0
